Make galaxy[version-incorrect] rule opt-in (#2103)
The docstring and description attribute for the GalaxyRule class have
been updated as to reflect the separation of galaxy[version-incorrect]
and that the rule does more than checking of a changelog.
