+++ docs/rules/galaxy-version-incorrect.md
@@ -1,0 +1,1 @@
+../../src/ansiblelint/rules/galaxy_version_incorrect.md
