+++ test/test_rules_collection.py
@@ -175,5 +175,5 @@ def test_rules_id_format(config_options: Options) -> None:
             rule.help or rule.description or rule.__doc__
         ), f"Rule {rule.id} must have at least one of:  .help, .description, .__doc__"
     assert "yaml" in keys, "yaml rule is missing"
-    assert len(rules) == 50  # update this number when adding new rules!
+    assert len(rules) == 51  # update this number when adding new rules!
     assert len(keys) == len(rules), "Duplicate rule ids?"
